Fix crash from multiple react-native-picker libraries
The app would crash if you navigate from Create Custom to Settings, or vice
versa. This was caused by two separate versions of @react-native-picker being
in use at the same time (one from react-native-picker-select, and the one that
I installed). Downgrading the version in package.json to match the version
in react-native-picker-select resolves this crash